requirements.txt
numpy~=1.21.6
tensorflow~=2.9.0
pillow~=9.2.0
tqdm~=4.64.1